+++ src/App.css
@@ -66,9 +66,11 @@
 button{
     background-image: url("CSS-Images/footer.jpg");
     color: white;
+    border: none;
 }
 
 .makeButton{
+    border: none;
     background-image: url("CSS-Images/footer.jpg");
     color: white;
     font-size: 25px;
@@ -137,8 +139,7 @@ label{
 }
 
 .scroll{
-    width:100%;
-    max-height: 100px;
+    max-height: min-content;
     overflow-y: scroll;
 }
 
@@ -149,6 +150,7 @@ label{
     text-align: center;
     caption-side: top;
 }
+
 table {
     border-collapse: collapse;
 }
@@ -161,6 +163,14 @@ th, td {
     display: none;
 }
 .remove{
-    width: 25%;
-    margin-left: 20px;
+    color: black;
+    background-color: #b63939;
+    width: auto;
+    height: auto;
+    border: none;
+    background-image: none;
+}
+.remove:hover{
+    color: white;
+    background-color: #8a2a2a;
 }
