+++ src/components/CreateCourse.tsx
@@ -1,0 +1,79 @@
+import React from "react";
+import "../App.css";
+import { Button, Form } from "react-bootstrap";
+
+type ChangeEvent = React.ChangeEvent<
+    HTMLTextAreaElement | HTMLInputElement | HTMLSelectElement
+>;
+
+export function CreateCourse({
+    visible,
+    changeVisibility,
+    courseID,
+    setID,
+    courseName,
+    setName,
+    courseDescription,
+    setDescription,
+    credits,
+    changeCredits,
+    addCourse
+}: {
+    visible: boolean;
+    changeVisibility: () => void;
+    courseID: string;
+    setID: (value: string) => void;
+    courseName: string;
+    setName: (value: string) => void;
+    courseDescription: string;
+    setDescription: (value: string) => void;
+    credits: number;
+    changeCredits: (value: string) => void;
+    addCourse: () => void;
+    resetState: () => void;
+}): JSX.Element {
+    return (
+        <div>
+            <Button onClick={changeVisibility}>
+                {visible ? "Hide Form" : "Create Course"}
+            </Button>
+            {visible && (
+                <div>
+                    <Form.Control
+                        type="string"
+                        value={courseID}
+                        onChange={(event: ChangeEvent) =>
+                            setID(event.target.value)
+                        }
+                        placeholder={"Enter Course ID"}
+                    />
+                    <Form.Control
+                        type="string"
+                        value={courseName}
+                        onChange={(event: ChangeEvent) =>
+                            setName(event.target.value)
+                        }
+                        placeholder={"Enter Course Name"}
+                    />
+                    <Form.Control
+                        type="string"
+                        value={courseDescription}
+                        onChange={(event: ChangeEvent) =>
+                            setDescription(event.target.value)
+                        }
+                        placeholder={"Enter Course Description"}
+                    />
+                    <label>Enter Number of Credits</label>
+                    <Form.Control
+                        type="number"
+                        value={credits}
+                        onChange={(event: ChangeEvent) =>
+                            changeCredits(event.target.value)
+                        }
+                    />
+                    <Button onClick={addCourse}>Submit Form</Button>
+                </div>
+            )}
+        </div>
+    );
+}
