+++ src/components/DropdownDegrees.tsx
@@ -1,12 +1,11 @@
 import React from "react";
-import { Degree } from "../interfaces/course-Degree-Semester";
+import { Degree } from "../interfaces/Course-Degree-Semester";
 import "../App.css";
 import { Form } from "react-bootstrap";
 
 type ChangeEvent = React.ChangeEvent<
     HTMLTextAreaElement | HTMLInputElement | HTMLSelectElement
 >;
-//export let currentSelectedDegree = AllDegrees[0];
 
 export function DropdownDegrees({
     degrees,
@@ -17,17 +16,6 @@ export function DropdownDegrees({
     degree: Degree;
     updateDegree: (event: ChangeEvent) => void;
 }): JSX.Element {
-    // const degrees = [...AllDegrees];
-    // const [degree, setDegree] = useState<Degree>(degrees[0]);
-    // function updateDegree(event: ChangeEvent) {
-    //     const newDegreeName = event.target.value;
-    //     const newDegree = degrees.find(
-    //         (degree: Degree): boolean => degree.name === newDegreeName
-    //     );
-    //     if (newDegree !== undefined) {
-    //         setDegree(newDegree);
-    //     }
-    // }
     return (
         <div>
             <Form.Group controlId="Degrees">
