+++ src/components/SchedulerPage.tsx
@@ -9,10 +9,9 @@ import {
     Season,
     SeasonsList,
     SemesterPlanner
-} from "../interfaces/course-Degree-Semester";
+} from "../interfaces/Course-Degree-Semester";
 import { ExportCSV } from "./ExportCSV";
-import { EditCourse } from "./EditCourse";
-//import { currentSelectedDegree } from "./DropdownDegrees";
+import { CoursesLists } from "./CourseLists";
 
 type ChangeEvent = React.ChangeEvent<
     HTMLTextAreaElement | HTMLInputElement | HTMLSelectElement
@@ -29,9 +28,6 @@ export function SchedulerPage({
 }): JSX.Element {
     //semester state
     const [showSemForm, setSemesterForm] = useState<boolean>(false);
-    // const [updateSemesterList, setSemesterList] = useState<SemesterPlanner[]>(
-    //     degree.SemesterList
-    // );
     const [semExistsError, setSemesterExists] = useState<boolean>(false);
     //update degree semester list
     function setSemesterList(newList: SemesterPlanner[]) {
@@ -101,20 +97,12 @@ export function SchedulerPage({
                 SemesterSeason: currSeason,
                 TotalCredits: 0
             };
-            //setSemesterList([...updateSemesterList, newSemester]);
             degree.SemesterList = [...degree.SemesterList, newSemester];
             updateDegree;
         }
     }
     function removeSemester(currYear: number, currSeason: Season) {
         updateDegree;
-        //console.log(year + " : " + season);
-        // setSemesterList(
-        //     updateSemesterList.filter(
-        //         (sem: SemesterPlanner): boolean =>
-        //             sem.SemesterSeason != currSeason || sem.year != currYear
-        //     )
-        // );
         degree.SemesterList = degree.SemesterList.filter(
             (sem: SemesterPlanner): boolean =>
                 sem.SemesterSeason != currSeason || sem.year != currYear
@@ -147,14 +135,6 @@ export function SchedulerPage({
                             {" " + degree.CreditsRequired}
                         </label>
                     </Col>
-                    {
-                        //<Col>
-                        //<label>Number of Credits Planned: </label>
-                        //</Col>
-                        //<Col>
-                        //<label>Number of Credits Unplanned: </label>
-                        //</Col>
-                    }
                 </Row>
             </div>
             <div>
@@ -167,9 +147,7 @@ export function SchedulerPage({
                             <div>
                                 <Button onClick={() => updateSemesterForm()}>
                                     {" "}
-                                    {showSemForm
-                                        ? "Hide Add Semester Form"
-                                        : "Show Add Semester Form"}
+                                    {showSemForm ? "Hide Form" : "Add Semester"}
                                 </Button>
                                 <Button
                                     className={"remove"}
@@ -240,151 +218,8 @@ export function SchedulerPage({
                     Back
                 </Button>
             </footer>
-            {console.log(degree)}
         </div>
     );
-}
-
-function CoursesLists({
-    degree,
-    updateDegree
-}: {
-    degree: Degree;
-    updateDegree: (event: ChangeEvent) => void;
-}): JSX.Element {
-    return (
-        <div>
-            <Container>
-                <Row>
-                    <Col>
-                        <EditCourse></EditCourse>
-                    </Col>
-                    <Col>
-                        <label>Courses Taken or Planned</label>
-                        <PrintDegreesLists
-                            taken={true}
-                            degree={degree}
-                            updateDegree={updateDegree}
-                        ></PrintDegreesLists>
-                    </Col>
-                    <Col>
-                        <label>Courses Not Taken or Planned </label>
-                        <PrintDegreesLists
-                            taken={false}
-                            degree={degree}
-                            updateDegree={updateDegree}
-                        ></PrintDegreesLists>
-                    </Col>
-                </Row>
-            </Container>
-        </div>
-    );
-}
-function PrintDegreesLists({
-    taken,
-    degree,
-    updateDegree
-}: {
-    taken: boolean;
-    degree: Degree;
-    updateDegree: (event: ChangeEvent) => void;
-}): JSX.Element {
-    // this is going to be where the courses are printed
-    const [currentDegree, setDegree] = useState<Degree>(degree);
-    // const [currentCourseName, setCurrentCourseName] = useState<string>(
-    //     degree.CoursesRequired[0].name
-    // );
-    const [printCourses, setPrintTakenOrNot] = useState<Course[]>(
-        currentDegree.CoursesRequired.filter(
-            (course: Course): boolean => course.taken === taken
-        )
-    );
-    const [progress, setProgress] = useState(0);
-    //const [currentTaken, setCurrentTaken] = useState<boolean>();
-
-    function updateList() {
-        updateDegree;
-        setDegree(degree);
-        // updating our list
-        setPrintTakenOrNot(
-            currentDegree.CoursesRequired.filter(
-                (course: Course): boolean => course.taken === taken
-            )
-        );
-    }
-
-    const scrollHandler = (event: React.UIEvent<HTMLDivElement>) => {
-        // this handles the scrolling of the box
-        const containerHeight = event.currentTarget.clientHeight;
-        const scrollHeight = event.currentTarget.scrollHeight;
-
-        const scrollTop = event.currentTarget.scrollTop;
-        setProgress(
-            progress + ((scrollTop + containerHeight) / scrollHeight) * 100
-        );
-        setDegree(degree);
-        // updating our list
-        setPrintTakenOrNot(
-            currentDegree.CoursesRequired.filter(
-                (course: Course): boolean => course.taken === taken
-            )
-        );
-    };
-
-    return (
-        <div>
-            <Button onClick={updateList}>Refresh {taken ? "✔️" : "❌"}</Button>
-            <div style={styles.container} onScroll={scrollHandler}>
-                {printCourses.map((currentCourse: Course) => (
-                    <div key={currentCourse.name}>
-                        Course ID: {currentCourse.courseID}
-                        {currentCourse.taken_String}
-                        <br></br>
-                        Pre Requisite:
-                        {currentCourse.prerecs.map(
-                            (currentPreRec: Course) =>
-                                currentPreRec.courseID +
-                                " " +
-                                currentPreRec.taken_String +
-                                "\n"
-                        )}
-                        <br></br>
-                        ____________
-                    </div>
-                ))}
-            </div>
-        </div>
-    );
-}
-// Styling
-const styles = {
-    container: {
-        width: 275,
-        height: 300,
-        margin: "30px auto",
-        overflowY: "auto",
-        overflowX: "hidden",
-        background: "mintcream"
-    },
-    list: {
-        width: "100%"
-    },
-    progressBar: {
-        width: 600,
-        height: 20,
-        margin: "auto",
-        backgroundColor: "#bbb"
-    },
-    progressValue: {
-        height: "100%",
-        backgroundColor: "steelblue"
-    },
-    text: {
-        textAlign: "center"
-    }
-} as const;
-{
-    /** Code for the scrolly box and the progress bar used from https://www.kindacode.com/article/react-typescript-handling-onscroll-event/*/
 }
 
 export default SchedulerPage;
